+++ app/[lng]/(main)/backup/layout.tsx
@@ -1,6 +1,3 @@
-import dynamic from 'next/dynamic'
-import { Box, Grid, Toolbar, Container } from '@mui/material'
-
 // Types
 import type { Metadata } from 'next'
 import type { LayoutProps } from '@/types/app/layouts'
@@ -11,9 +8,6 @@ import LocaleConfig from '@/config/locale'
 
 // Helpers
 import { useTranslation } from '@/helpers/i18n/server'
-
-// Components
-const AppbarOrganism = dynamic(() => import('@/components/organisms/appbar'))
 
 // Metadata
 export const generateMetadata = async (props: PageProps): Promise<Metadata> => {
@@ -27,7 +21,7 @@ export const generateMetadata = async (props: PageProps): Promise<Metadata> => {
 	// eslint-disable-next-line react-hooks/rules-of-hooks
 	const { t } = await useTranslation(lng)
 
-	const metadata: Metadata = { title: t('links:home') }
+	const metadata: Metadata = { title: t('links:backup') }
 
 	return metadata
 }
@@ -39,29 +33,7 @@ const MainLayout = (props: LayoutProps) => {
 	// Variables
 	const lng = params?.lng ?? LocaleConfig.default
 
-	return (
-		<Box display="flex">
-			<AppbarOrganism lng={lng} />
-
-			<Grid container>
-				<Container
-					maxWidth="md"
-					sx={{
-						textAlign: 'center',
-						zIndex: 1
-					}}
-				>
-					<Toolbar
-						sx={{
-							boxSizing: 'content-box',
-							py: { md: 2 }
-						}}
-					/>
-					{children}
-				</Container>
-			</Grid>
-		</Box>
-	)
+	return <>{children}</>
 }
 
 export default MainLayout