+++ components/templates/backup.tsx
@@ -1,0 +1,13 @@
+'use client'
+
+// Types
+import type { HomeProps } from '@/types/components/templates/home'
+
+const BackupTemplate = (props: HomeProps) => {
+	// Props
+	const { lng } = props
+
+	return <>Backup</>
+}
+
+export default BackupTemplate