+++ components/templates/home.tsx
@@ -1,7 +1,6 @@
 'use client'
 
 import dynamic from 'next/dynamic'
-import { Grid, Toolbar, Container } from '@mui/material'
 
 // Types
 import type { HomeProps } from '@/types/components/templates/home'
@@ -14,22 +13,7 @@ const HomeTemplate = (props: HomeProps) => {
 	// Props
 	const { lng } = props
 
-	return (
-		<Grid container>
-			<Container
-				maxWidth="md"
-				sx={{
-					mt: { xs: 2, md: 10 },
-					textAlign: 'center',
-					zIndex: 1
-				}}
-			>
-				<Toolbar />
-
-				{lng === 'fa' ? <HomeFa /> : <HomeEn />}
-			</Container>
-		</Grid>
-	)
+	return lng === 'fa' ? <HomeFa /> : <HomeEn />
 }
 
 export default HomeTemplate