+++ components/templates/settings.tsx
@@ -1,0 +1,13 @@
+'use client'
+
+// Types
+import type { HomeProps } from '@/types/components/templates/home'
+
+const SettingsTemplate = (props: HomeProps) => {
+	// Props
+	const { lng } = props
+
+	return <>Settings</>
+}
+
+export default SettingsTemplate