+++ types/components/atoms/buttons/icon.d.ts
@@ -1,7 +1,9 @@
 // Types
 import type { ReactNode } from 'react'
+import type { Color } from '@mui/material'
 
 export type IconButtonProps = {
+	color: Color
 	icon: ReactNode
 	onClick?: () => void
 }